Fix some flaky enrionment tests
This PR attempts to address two flaky environment tests

One is how we attempt to read environment variables from a child process
on windows. For this, I updated the code original referenced to an
updated version from the author.

The second is a refactor that to fix flaky tests that rely on
Environment since they run in parallel and appear to collide
occasionally