SelfUpdater: Add `--no-same-owner` to `tar` invocation
This avoids having `tar` attempt to set the original owner and access times for files when extracting them as `root`.  While not best practice on multi-user systems, single-user systems (such as Alpine, or a chroot user namespace) run all processes as `root`, and it is possible to get a `tar` on that system that will attempt to set e.g. `UID 1000` from a tarball when that doesn't make any sense.  As the Azure Pipelines agent doesn't really care about file ownership in this case, this flag should be completely safe to set, and should have no effect outside of the cases where it is needed.