Default System.PreferGitFromPath to false
Prior to the port to AgentKnobs, we used git
from the path always on Linux and Mac, but on
Windows we preferred the MinGit shipped with the
agent, unless the System.PreferGitFromPath variable
was present and true. The fact that this config
defaulted to false wasn't obvious because it
relied on Boolean.TryParse setting the output
variable to false on a failure parsing a non-existent
environment variable. The default has changed by
mistake in the port to knobs, causing checkouts to
fail on Windows systems with no git in the path.