Bugfixing localization problems in several UnitTests or at least mentioning the possible reason for them failing on non-english systems in the failing message
Also bugfixing confliciting assignment of what the primary repository is. This is always be the self repository for the RepositoryUtil.IsPrimaryRepositoryName method, but was always the first one checked out for the RepositoryUtil.GetPrimaryRepository method, even if that wasn't the self repository. This could for example result in the wrong repositoryUrl in the trackingConfig/SourceFolder.json. Now it's the self repository that get's marked as the primary repository for the RepositoryUtil.GetPrimaryRepository method too, unless no self repository was checked out, then it will be the first repository that gets checked out. So now the result of both methods should be consistent.