[Checkout] Configure sparse checkouts before fetch
This addresses two issues - first, it allows us to avoid fetching files
when checking out a clean repo, as if we pre-configure sparse checkouts
we won't fetch unneeded trees and blobs. Second, it allows us to avoid
not fetching files incorrectly if we're moving a checkout from sparse to
non-sparse.

Fixes #5091